fix bugs and move raw loader to deperacated files to provide support for users store data only on rank 0 machine
